dekita

+++ app/Http/Controllers/TasksController.php
@@ -14,11 +14,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-         $tasks = Task::all();
+         $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -48,10 +56,9 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-         $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,'status'=> $request->status,
+        ]);
 
         return redirect('/');
     }
@@ -65,10 +72,12 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = task::find($id);
-
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+        return redirect('/');
     }
 
 
@@ -81,10 +90,12 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = task::find($id);
-
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        return redirect('/');
     }
 
     /**
@@ -101,10 +112,11 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         $task = Task::find($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        
-        $task->save();
+        if (\Auth::id() === $task->user_id){
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+        }
 
         return redirect('/');
     }
@@ -117,9 +129,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::find($id);
-        $task->delete();
-
+        $task = Task::find($id);
+        
+        if (\Auth::id() === $task->user_id){
+            $task->delete();
+        }
         return redirect('/');
     }
 }
